requirements.txt
Flask
requests
colorama
python-dotenv
huaweicloudsdkcore>=3.0.0
huaweicloudsdkcph>=3.0.0